add cache tag invalidation support for InMemoryLRUCache
