--- packages/apollo-server-caching/src/__tests__/InMemoryKeyValueCache.test.ts
@@ -1,3 +1,0 @@
-import { testKeyValueCache } from '../tests';
-import { InMemoryKeyValueCache } from '../InMemoryKeyValueCache';
-testKeyValueCache(new InMemoryKeyValueCache());
