+++ packages/apollo-server-caching/src/__tests__/InMemoryLRUCache.test.ts
@@ -1,0 +1,3 @@
+import { testKeyValueCache } from '../testsuite';
+import { InMemoryLRUCache } from '../InMemoryLRUCache';
+testKeyValueCache(new InMemoryLRUCache());
